requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
alembic==1.7.5
argon2-cffi==21.1.0
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
bleach==4.1.0
certifi==2021.10.8
certipy==0.1.3
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.7
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.9
django-appconf==1.0.5
django-ckeditor==6.1.0
django-cors-middleware==1.5.0
django-crispy-forms
django-debug-toolbar==3.2.2
django-filter==21.1
django-geoip==0.5.2.1
django-js-asset==1.2.2
django-maintenance-mode==0.16.1
django-mptt==0.13.4
django-multiselectfield==0.1.12
django-oauth-toolkit==1.5.0
django-user-agents==0.4.0
django-user-sessions==1.7.1
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
entrypoints==0.3
frozenlist==1.2.0
geoip2==4.4.0
greenlet==1.1.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
incremental==21.3.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-telemetry==0.1.0
jupyterhub==2.0.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
jwcrypto==1.0
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
maxminddb==2.2.0
mistune==0.8.4
msgpack==1.0.2
multidict==5.2.0
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
oauthenticator==14.2.0
oauthlib==3.1.1
packaging==21.3
pamela==1.0.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip-chill==1.0.1
prometheus-client==0.12.0
prompt-toolkit==3.0.23
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-json-logger==2.0.2
python-memcached==1.59
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
SQLAlchemy==1.4.27
sqlparse==0.4.2
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.0.0
ua-parser==0.10.0
urllib3==1.26.7
user-agents==2.2.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wrapt==1.13.3
yarl==1.7.2
zipp==3.6.0
zope.interface==5.4.0
